Moved "Mouse" class into seperate file.
